pyproject.toml
[tool.poetry]
name = "geneus-loci"
version = "0.1.0"
description = ""
authors = ["Yongha Hwang <yongha.hwang@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
pandas = "1.3"
geopandas = "^0.10.2"
PyYAML = "^6.0"
GDAL = "3.3.2"
"ruamel.yaml" = "^0.17.21"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
meta = 'geneus_loci.meta:main'
convert = 'geneus_loci.convert:main'
filter = 'geneus_loci.convert:filter'
